+++ CHANGELOG.md
@@ -26,6 +26,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Fixed `callVariant` of 'no reference out node found'. #842
 
+- Fixed `mergeFasta` to remove redundant FASTA header entries. #846
+
 ## [1.2.1] - 2023-10-05
 
 ### Add